+++ lib/display_identify_overlay.dart
@@ -1,5 +1,6 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/core/platform_detector.dart';
+export 'src/display_identify_overlay.dart';
+export 'src/models/monitor_info.dart';
+export 'src/models/overlay_options.dart';
+export 'src/models/overlay_style.dart';
+export 'src/exceptions/display_identify_exception.dart';